debug.py bug fixed

+++ debug.py
@@ -154,7 +154,7 @@ def main(input_filename, model_filename, input_size, output_filename):
         
         if output_filename is not None:
             with open(output_filename, 'wb') as f:
-                np.array(len(save_output.outputs)).tofile(f)
+                np.array(len(save_output.outputs), dtype=np.int32).tofile(f)
                 for o in save_output.outputs:
                     f.write(o[0].encode())
                     np.array(o[1], dtype=np.int32).tofile(f)
